requirements.txt
numpy==1.17.0
scipy==1.3.0
matplotlib==3.1.1
pandas==0.25.0
scikit-learn==0.21.3
nilearn==0.5.2
nibabel==2.5.0
seaborn==0.9.0
attrs==19.1.0
